+++ src/Configuration/Permissions/PermissionInitializer.php
@@ -6,13 +6,17 @@ use DemosEurope\DemosplanAddon\Contracts\Config\GlobalConfigInterface;
 use DemosEurope\DemosplanAddon\Permission\PermissionConditionBuilder;
 use DemosEurope\DemosplanAddon\Permission\PermissionInitializerInterface;
 use DemosEurope\DemosplanAddon\Permission\ResolvablePermissionCollectionInterface;
+use DemosEurope\DemosplanAddon\XBeteiligung\Enum\ProcedureMessageTyp;
+use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 
 class PermissionInitializer implements PermissionInitializerInterface
 {
     private bool $restrictedAccess;
-    public function __construct(GlobalConfigInterface $globalConfig)
+    private string $procedureMessageType;
+    public function __construct(GlobalConfigInterface $globalConfig, ParameterBagInterface $parameterBag)
     {
         $this->restrictedAccess = $globalConfig->hasProcedureUserRestrictedAccess();
+        $this->procedureMessageType = $parameterBag->get('procedure_message_type');
     }
 
     /**
@@ -25,5 +29,41 @@ class PermissionInitializer implements PermissionInitializerInterface
             PermissionConditionBuilder::start()
                 ->enableAlways()
         );
+
+        if (ProcedureMessageTyp::KOMMUNAL->value === $this->procedureMessageType)
+        {
+            $permissionCollection->configurePermissionInstance(
+                Features::feature_procedure_message_kom_create(),
+                PermissionConditionBuilder::start()->enableAlways()
+            );
+
+            $permissionCollection->configurePermissionInstance(
+                Features::feature_procedure_message_kom_update(),
+                PermissionConditionBuilder::start()->enableAlways()
+            );
+
+            $permissionCollection->configurePermissionInstance(
+                Features::feature_procedure_message_kom_delete(),
+                PermissionConditionBuilder::start()->enableAlways()
+            );
+        }
+
+        if (ProcedureMessageTyp::RAUMORDNUNG->value === $this->procedureMessageType)
+        {
+            $permissionCollection->configurePermissionInstance(
+                Features::feature_procedure_message_rog_create(),
+                PermissionConditionBuilder::start()->enableAlways()
+            );
+
+            $permissionCollection->configurePermissionInstance(
+                Features::feature_procedure_message_rog_update(),
+                PermissionConditionBuilder::start()->enableAlways()
+            );
+
+            $permissionCollection->configurePermissionInstance(
+                Features::feature_procedure_message_rog_delete(),
+                PermissionConditionBuilder::start()->enableAlways()
+            );
+        }
     }
 }