+++ src/EventSubscriber/XBeteiligungEventSubscriber.php
@@ -10,7 +10,10 @@
 
 namespace DemosEurope\DemosplanAddon\XBeteiligung\EventSubscriber;
 
+use DemosEurope\DemosplanAddon\Contracts\Entities\ProcedureInterface;
 use DemosEurope\DemosplanAddon\Contracts\Events\PostNewProcedureCreatedEventInterface;
+use DemosEurope\DemosplanAddon\Permission\PermissionEvaluatorInterface;
+use DemosEurope\DemosplanAddon\XBeteiligung\Configuration\Permissions\Features;
 use DemosEurope\DemosplanAddon\XBeteiligung\Debugger\XBeteiligungDebugger;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\XBeteiligungService;
 use Exception;
@@ -19,8 +22,9 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 class XBeteiligungEventSubscriber implements EventSubscriberInterface
 {
     public function __construct(
-        private readonly XBeteiligungService $xBeteiligungService,
-        private readonly XBeteiligungDebugger $xBeteiligungDebugger
+        private readonly PermissionEvaluatorInterface $permissionEvaluator,
+        private readonly XBeteiligungDebugger $xBeteiligungDebugger,
+        private readonly XBeteiligungService $xBeteiligungService
     ) {
     }
 
@@ -39,11 +43,23 @@ class XBeteiligungEventSubscriber implements EventSubscriberInterface
      */
     public function newProcedureCreated(PostNewProcedureCreatedEventInterface $event): void
     {
-        $xml = $this->xBeteiligungService->createProcedureNew401FromObject($event->getProcedure());
-        $procedureMessage = $this->xBeteiligungService->createProcedureMessage($xml, $event->getProcedure()->getId());
+        if ($this->permissionEvaluator->isPermissionEnabled(Features::feature_procedure_message_kom_create())) {
+            $xml = $this->xBeteiligungService->createProcedureNew401FromObject($event->getProcedure());
+            $this->createProcedureMessage($xml, $event->getProcedure());
+        }
+
+        if ($this->permissionEvaluator->isPermissionEnabled(Features::feature_procedure_message_rog_create())) {
+            $xml = $this->xBeteiligungService->createXMLFor301($event->getProcedure());
+            $this->createProcedureMessage($xml, $event->getProcedure());
+        }
+    }
+
+    private function createProcedureMessage(string $xml, ProcedureInterface $procedure): void
+    {
+        $procedureMessage = $this->xBeteiligungService->createProcedureMessage($xml, $procedure->getId());
         $this->xBeteiligungService->saveProcedureMessage($procedureMessage);
         $this->xBeteiligungDebugger->createDebugMessageForCreatedXML(
-            $event->getProcedure(),
+            $procedure,
             $xml,
             'created'
         );